refactor(runtime): extract run fact usage parser
